requirements.txt
coloredlogs==15.0.1
fastapi==0.70.0
PyYAML==5.4.1
starlette==0.16.0
